# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443/
  description: Generated server url
paths:
...
  /api/users/{id}/image:
    get:
      tags:
      - user-rest-controller
      operationId: getImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - user-rest-controller
      operationId: putUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: uri
...
  /api/portfolioItems/users/{userId}:
    get:
      tags:
      - portfolio-item-rest-controller
      operationId: getPortfolioItems
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagePortfolioItem'
    post:
      tags:
      - portfolio-item-rest-controller
      operationId: postPortfolioItem
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortfolioItemDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PortfolioItem'
...
  /api/portfolioItems/{portfolioItemId}/previewImage:
    get:
      tags:
      - portfolio-item-rest-controller
      operationId: getPreviewImage
      parameters:
      - name: portfolioItemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/portfolioItems/{portfolioItemId}/image3:
    get:
      tags:
      - portfolio-item-rest-controller
      operationId: getImage3
      parameters:
      - name: portfolioItemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/portfolioItems/{portfolioItemId}/image2:
    get:
      tags:
      - portfolio-item-rest-controller
      operationId: getImage2
      parameters:
      - name: portfolioItemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/portfolioItems/{portfolioItemId}/image1:
    get:
      tags:
      - portfolio-item-rest-controller
      operationId: getImage1
      parameters:
      - name: portfolioItemId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        surname:
          type: string
        bornDate:
          type: string
          format: date-time
        email:
          type: string
        password:
          type: string
        phoneNumber:
          type: string
        website:
          type: string
        city:
          type: string
        category:
          type: string
        degree:
          type: string
        freelance:
          type: string
        description:
          type: string
        profilePhoto:
          type: string
          format: binary
...
    PortfolioItemDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: string
        name:
          type: string
        description:
          type: string
        previewImg:
          type: string
          format: binary
        image1:
          type: string
          format: binary
        image2:
          type: string
          format: binary
        image3:
          type: string
          format: binary
        category:
          type: string
        client:
          type: string
        url:
          type: string
        date:
          type: string
          format: date-time
    PortfolioItem:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        client:
          type: string
        url:
          type: string
        date:
          type: string
          format: date-time
...
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
    PagePortfolioItem:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioItem'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean